requirements.txt
# Content Orchestrator Dependencies

# Data validation
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# API interactions with DeepSeek
openai>=1.0.0,<2.0.0

# Configuration management
python-dotenv>=1.0.0,<2.0.0

# Web UI
gradio>=4.0.0,<5.0.0

# Video transcript extraction
yt-dlp>=2024.0.0,<2025.0.0